development changes

+++ actions/project-group.ts
@@ -94,13 +94,45 @@ export const upsertPg = action
     const { code, pics, ...data } = parsedInput
     const { userId } = ctx
 
+    const include: Prisma.ProjectGroupInclude = {
+      projectGroupPics: {
+        //* only return the pic that has role 'admin' or which they allowed to 'receive notifications (owner)' permission action
+        where: {
+          user: {
+            OR: [
+              {
+                role: {
+                  rolePermissions: {
+                    some: {
+                      permission: { code: PERMISSIONS_CODES['PROJECT GROUPS'] },
+                      actions: { has: PERMISSIONS_ALLOWED_ACTIONS.RECEIVE_NOTIFICATIONS_OWNER },
+                    },
+                  },
+                },
+              },
+              {
+                role: {
+                  key: 'admin',
+                },
+              },
+            ],
+          },
+        },
+        select: { userCode: true },
+      },
+    }
+
     try {
       if (code !== -1) {
+        const existingPg = await db.projectGroup.findUnique({ where: { code }, include })
+
+        if (!existingPg) {
+          return { error: true, status: 404, message: 'Project group not found!', action: 'UPSERT_PROJECT_GROUP' }
+        }
+
         const [updatedPg] = await db.$transaction([
-          //* update project group
           db.projectGroup.update({ where: { code }, data: { ...data, updatedBy: userId } }),
 
-          //* delete existing project group pics
           db.projectGroupPic.deleteMany({ where: { projectGroupCode: code } }),
 
           //* create new project group pics
@@ -108,6 +140,64 @@ export const upsertPg = action
             data: pics.map((p) => ({ projectGroupCode: code, userCode: p })),
           }),
         ])
+
+        //* create notifications
+        // void createNotification(ctx, {
+        //   permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+        //   title: 'Project Group Updated',
+        //   message: `A project group (#${updatedPg.code}) was updated by ${ctx.fullName}.`,
+        //   link: `/project/groups/${updatedPg.code}/view`,
+        //   entityType: 'ProjectGroup' as Prisma.ModelName,
+        //   entityCode: updatedPg.code,
+        //   entityId: updatedPg.id,
+        //   userCodes: [],
+        // }).then((data) => {
+        //   if (data.error) return
+
+        //   const currentPics = existingPg.projectGroupPics.map((pgp) => pgp.userCode)
+        //   const newlyAssignedPics = pics.filter((pic) => !currentPics.includes(pic))
+        //   const unAssignedPics = currentPics.filter((pic) => !pics.includes(pic))
+
+        //   //* if the user is assigned or unassigned as a pic, then skip notify will be false to notify the user otherwise true
+        //   const isNewlyAssignedPicsIsSkipNotify = newlyAssignedPics.includes(ctx.userCode)
+        //   const isUnassignedPicsIsSkipNotify = unAssignedPics.includes(ctx.userCode)
+
+        //   //* create notification for newly assigned pics
+        //   if (newlyAssignedPics.length > 0) {
+        //     void createNotification(
+        //       ctx,
+        //       {
+        //         permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+        //         title: 'Project Group PIC Assigned',
+        //         message: `You were assigned as a PIC to project group (#${updatedPg.code}) by ${ctx.fullName}.`,
+        //         link: `/project/groups/${updatedPg.code}/view`,
+        //         entityType: 'ProjectGroup' as Prisma.ModelName,
+        //         entityCode: updatedPg.code,
+        //         entityId: updatedPg.id,
+        //         userCodes: newlyAssignedPics,
+        //       },
+        //       { skipNotify: isNewlyAssignedPicsIsSkipNotify ? false : true }
+        //     )
+        //   }
+
+        //   //* create notification for unassigned pics
+        //   if (unAssignedPics.length > 0) {
+        //     void createNotification(
+        //       ctx,
+        //       {
+        //         permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+        //         title: 'Project Group PIC Unassigned',
+        //         message: `You were unassigned as a PIC to project group (#${updatedPg.code}) by ${ctx.fullName}.`,
+        //         link: `/project/groups`,
+        //         entityType: 'ProjectGroup' as Prisma.ModelName,
+        //         entityCode: updatedPg.code,
+        //         entityId: updatedPg.id,
+        //         userCodes: unAssignedPics,
+        //       },
+        //       { skipNotify: isUnassignedPicsIsSkipNotify ? false : true }
+        //     )
+        //   }
+        // })
 
         return {
           status: 200,
@@ -129,24 +219,7 @@ export const upsertPg = action
             },
           },
         },
-        include: {
-          projectGroupPics: {
-            //* only return the p.i.c that has role which they allowed to 'receive notifications (owner)' permission action
-            where: {
-              user: {
-                role: {
-                  rolePermissions: {
-                    some: {
-                      permission: { code: PERMISSIONS_CODES['PROJECT GROUPS'] },
-                      actions: { has: PERMISSIONS_ALLOWED_ACTIONS.RECEIVE_NOTIFICATIONS_OWNER },
-                    },
-                  },
-                },
-              },
-            },
-            select: { userCode: true },
-          },
-        },
+        include,
       })
 
       // //* create notification
@@ -159,13 +232,34 @@ export const upsertPg = action
       //   entityCode: newPg.code,
       //   entityId: newPg.id,
       //   userCodes: [],
+      // }).then((data) => {
+      //   if (data.error) return
+
+      //   //* only notify when notification of created project group was successful
+      //   //* notify assigned pics
+      //   const assignedPics = newPg.projectGroupPics.map((pgp) => pgp.userCode)
+
+      //   //* if the user is assigned as a pic, then skip notify will be false to notify the user otherwise true
+      //   const isAssignedPics = assignedPics.includes(ctx.userCode)
+
+      //   //* create notification for assigned pics
+      //   if (assignedPics.length > 0) {
+      //     void createNotification(
+      //       ctx,
+      //       {
+      //         permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //         title: 'Project Group PIC Assigned',
+      //         message: `You were assigned as a PIC to project group (#${newPg.code}) by ${ctx.fullName}.`,
+      //         link: `/project/groups/${newPg.code}/view`,
+      //         entityType: 'ProjectGroup' as Prisma.ModelName,
+      //         entityCode: newPg.code,
+      //         entityId: newPg.id,
+      //         userCodes: assignedPics,
+      //       },
+      //       { skipNotify: isAssignedPics ? false : true }
+      //     )
+      //   }
       // })
-      //   .then(() => {
-      //     //* notify assigned p.i.cs
-      //     //* do not notify already notified user
-      //     // Promise.all()
-      //   })
-      //   .catch((error) => console.error(error))
 
       return {
         status: 200,
@@ -196,6 +290,18 @@ export const deleletePg = action
 
       await db.projectGroup.update({ where: { code: data.code }, data: { deletedAt: new Date(), deletedBy: ctx.userId } })
 
+      //* create notification
+      // void createNotification(ctx, {
+      //   permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //   title: 'Project Group Deleted',
+      //   message: `A project group (#${projectGroup.code}) was deleted by ${ctx.fullName}.`,
+      //   link: `/project/groups/${projectGroup.code}/view`,
+      //   entityType: 'ProjectGroup' as Prisma.ModelName,
+      //   entityCode: projectGroup.code,
+      //   entityId: projectGroup.id,
+      //   userCodes: [],
+      // })
+
       return { status: 200, message: 'Project group deleted successfully!', action: 'DELETE_PROJECT_GROUP' }
     } catch (error) {
       console.error(error)
@@ -212,13 +318,25 @@ export const deleletePg = action
 export const restorePg = action
   .use(authenticationMiddleware)
   .schema(paramsSchema)
-  .action(async ({ parsedInput: data }) => {
+  .action(async ({ ctx, parsedInput: data }) => {
     try {
       const projectGroup = await db.projectGroup.findUnique({ where: { code: data.code } })
 
       if (!projectGroup) return { error: true, status: 404, message: 'Project group not found!', action: 'RESTORE_PROJECT_GROUP' }
 
       await db.projectGroup.update({ where: { code: data.code }, data: { deletedAt: null, deletedBy: null } })
+
+      //* create notification
+      // void createNotification(ctx, {
+      //   permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //   title: 'Project Group Restored',
+      //   message: `A project group (#${projectGroup.code}) was restored by ${ctx.fullName}.`,
+      //   link: `/project/groups/${projectGroup.code}/view`,
+      //   entityType: 'ProjectGroup' as Prisma.ModelName,
+      //   entityCode: projectGroup.code,
+      //   entityId: projectGroup.id,
+      //   userCodes: [],
+      // })
 
       return { status: 200, message: 'Project group retored successfully!', action: 'RESTORE_PROJECT_GROUP' }
     } catch (error) {
@@ -283,6 +401,18 @@ export const importPgs = action
         status: progress >= 100 || isLastRow ? 'completed' : 'processing',
       }
 
+      // if (updatedStats.status === 'completed') {
+      //   //* create notification
+      //   void createNotification(ctx, {
+      //     permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //     title: 'Project Group Imported',
+      //     message: `New project group${total > 1 ? 's were' : ' was'} imported by ${ctx.fullName}.`,
+      //     link: `/project/groups`,
+      //     entityType: 'ProjectGroup' as Prisma.ModelName,
+      //     userCodes: [],
+      //   })
+      // }
+
       return {
         status: 200,
         message: `${updatedStats.completed} project group created successfully!`,
@@ -318,10 +448,38 @@ export const updatePgPics = action
     const { code, pics } = parsedInput
     const { userId } = ctx
 
-    try {
-      const projectGroup = await db.projectGroup.findUnique({ where: { code } })
+    const include: Prisma.ProjectGroupInclude = {
+      projectGroupPics: {
+        //* only return the pic that has role which they allowed to 'receive notifications (owner)' permission action
+        where: {
+          user: {
+            OR: [
+              {
+                role: {
+                  rolePermissions: {
+                    some: {
+                      permission: { code: PERMISSIONS_CODES['PROJECT GROUPS'] },
+                      actions: { has: PERMISSIONS_ALLOWED_ACTIONS.RECEIVE_NOTIFICATIONS_OWNER },
+                    },
+                  },
+                },
+              },
+              {
+                role: {
+                  key: 'admin',
+                },
+              },
+            ],
+          },
+        },
+        select: { userCode: true },
+      },
+    }
 
-      if (!projectGroup) {
+    try {
+      const pg = await db.projectGroup.findUnique({ where: { code }, include })
+
+      if (!pg) {
         return { error: true, status: 404, message: 'Project group not found!', action: 'UPDATE_PROJECT_GROUP_PICS' }
       }
 
@@ -342,9 +500,63 @@ export const updatePgPics = action
         }),
       ])
 
+      //* create notifications
+      // void createNotification(ctx, {
+      //   permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //   title: 'Project Group Updated',
+      //   message: `A project group (#${pg.code}) was updated by ${ctx.fullName}.`,
+      //   link: `/project/groups/${pg.code}/view`,
+      //   entityType: 'ProjectGroup' as Prisma.ModelName,
+      //   entityCode: pg.code,
+      //   entityId: pg.id,
+      //   userCodes: [],
+      // }).then((data) => {
+      //   if (data.error) return
+
+      //   const currentPics = pg.projectGroupPics.map((pgp) => pgp.userCode)
+      //   const newlyAssignedPics = pics.filter((pic) => !currentPics.includes(pic))
+      //   const unAssignedPics = currentPics.filter((pic) => !pics.includes(pic))
+
+      //   //* create notification for newly assigned pics
+      //   if (newlyAssignedPics.length > 0) {
+      //     void createNotification(
+      //       ctx,
+      //       {
+      //         permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //         title: 'Project Group PIC Assigned',
+      //         message: `You were assigned as a PIC to project group (#${pg.code}) by ${ctx.fullName}.`,
+      //         link: `/project/groups/${pg.code}/view`,
+      //         entityType: 'ProjectGroup' as Prisma.ModelName,
+      //         entityCode: pg.code,
+      //         entityId: pg.id,
+      //         userCodes: newlyAssignedPics,
+      //       },
+      //       { skipNotify: true }
+      //     )
+      //   }
+
+      //   //* create notification for unassigned pics
+      //   if (unAssignedPics.length > 0) {
+      //     void createNotification(
+      //       ctx,
+      //       {
+      //         permissionCode: PERMISSIONS_CODES['PROJECT GROUPS'],
+      //         title: 'Project Group PIC Unassigned',
+      //         message: `You were unassigned as a PIC to project group (#${pg.code}) by ${ctx.fullName}.`,
+      //         link: `/project/groups`,
+      //         entityType: 'ProjectGroup' as Prisma.ModelName,
+      //         entityCode: pg.code,
+      //         entityId: pg.id,
+      //         userCodes: unAssignedPics,
+      //       },
+      //       { skipNotify: true }
+      //     )
+      //   }
+      // })
+
       return {
         status: 200,
-        message: `Project group's P.I.Cs updated successfully!`,
+        message: `Project group's PICs updated successfully!`,
         action: 'UPDATE_PROJECT_GROUP_PICS',
         data: { projectGroup: updatedProjectGprojectGroup },
       }
